Allow loading from `custom_component` folder
Letting users load the component by placing the source file into
their `custom_components` directory helps people use it. That required
the following changes:

1. `import` the daikinskyport integration without requiring extra steps

Move the integrations `__init__.py` file into the source code
folder and import it through the local path. This choice preserves the
seperation of component and integration though different files. If a
proper integration project is eventually pulled out, like
python-ecobee-api, it can take the `daikinskyport.py` file without
pulling it out of other code. Try to follow
https://developers.home-assistant.io/docs/creating_component_code_review/#4-communication-with-devicesservices
as close as possible without actually making a seperate project right
now.

2. Break the circular dependencies of platforms requiring the component

When the component is loading it would attempt to load it`s
platforms (`sensor.py`,etc), but those would attempt to import the
component causing a failure. The platforms did this to attempt to get
the data object to allowing refreshing. Home Assistant guidelines say
to do this with the `hass` object.
https://developers.home-assistant.io/docs/creating_component_code_review/#3-componentplatform-communication

Add `const.py` to share common constants between the component and
platforms.

Remove `request_configuration`. It was not being used and prevented
using the `hass` object. The original ecobee code this was copied from
would call it during setup. That was not done for this component, or
currently needed. Future configuration work should be inspired by
https://developers.home-assistant.io/docs/config_entries_config_flow_handler