Invert the logic for when to record a value for the fault code sensor
There are conditions where the fault code sensor will return 255 to
indicate some sort of issue with that component (learned by experience).
With the old logic, we would return 'unknown'. With the new logic, no
value is returned at all, effectively leaving the sensor with whatever
state it previously had. This aligns with the same behavior we have for
sensors returning values such as 65535.

I'm not sure if ignoring 255 and not exposing that up to HA makes sense,
but I'm erring on the side of aligning behaviors between sensor types
within this component.